requirements.txt
streamlit>=1.38.0
groq>=0.9.0
sentence-transformers>=3.0.0
pinecone>=5.0.0
pypdf2>=3.0.1
python-docx>=1.1.2
python-dotenv>=1.0.1
pandas>=2.2.2
plotly>=5.22.0
langchain-text-splitters>=0.3.0